==> 06_COX_with_lags.ipynb <==
# UKB MODELS

!pip install statsmodels
!pip install lifelines==0.26.4

# Imports here.
import numpy as np
import pandas as pd
import os
import statsmodels.api as sm
import statsmodels.formula.api as smf
from statsmodels.stats.multitest import fdrcorrection
from lifelines import CoxPHFitter

import warnings
warnings.filterwarnings("ignore")

!dx download -r 'data/files_for_cox'

#Set variables
STUDY_ENDS = '2023-09-30'
STUDY_START = '1999-01-01'

#Get list of codes from prep notebook
ndd_list = ['AD', 'DEM', 'PD', 'VAS', 'MS', 'ALS']
codes = ['G47_DATE', 'F51_DATE', 'G47.3_DATE']
lags = ['0', '0-1', '1-5', '5-10', '10-15', '5-15']

#Get list of people with Welsh data -- datafield 26426
w = pd.read_csv('files_for_cox/Wales_dep_n21182_participant.csv')
wales_list = list(w['eid'])

timeline = '1999-2023'
model = 'COX'

results = []

for lag in lags:

    for ndd in ndd_list:

        #Load df
        df = pd.read_csv(f'files_for_cox/{ndd}_with_tenure_lags_45.csv', parse_dates = True)
 
        #Include this line in to exclude Welsh data
        #df = df[~df['ID'].isin(wales_list)]

        #Only selected codes with data
        codes_with_data = []
        for code in codes:
            m = df[['TOWNSEND', 'AGE_OF_RECRUIT', 'GENETIC_SEX', 'tenure', ndd, f'QC{lag}_' +code]]
            n=sum(m[f'QC{lag}_'+code])
            df_pair = m[m[f'QC{lag}_'+code]==1]
            n_pairs = sum(df_pair[ndd])
            if n == 0:
                pass
            elif n_pairs < 5:
                pass
            elif n == n_pairs:
                covariate = code
                HR = np.nan
                ci_min = np.nan
                ci_max = np.nan
                p = np.nan
                print(covariate, ndd, HR, ci_min, ci_max, p, n_pairs, n)
                results.append((covariate, ndd, model, timeline, lag, HR, ci_min, ci_max, p, n_pairs, n))
                pass
            else:
                codes_with_data.append(code)

        for code in codes_with_data:
            m = df[['TOWNSEND', 'AGE_OF_RECRUIT', 'GENETIC_SEX', 'tenure', ndd, f'QC{lag}_' +code]]
            n=sum(m[f'QC{lag}_'+code])
            df_pair = m[m[f'QC{lag}_'+code]==1]
            n_pairs = sum(df_pair[ndd])

            cph = CoxPHFitter()
            cph.fit(m, duration_col = 'tenure', event_col = ndd, show_progress=False, step_size = 0.1)
            #cph.print_summary()
            #cph.plot()

            actual_p = cph._compute_p_values()
            results_df = cph.summary
            results_df = results_df.reset_index()
            test = results_df.iloc[3]

            covariate = code
            HR = test['exp(coef)']
            ci_min = test['exp(coef) lower 95%']
            ci_max = test['exp(coef) upper 95%']
            p = actual_p[3]

            print(covariate, ndd, lag, HR, ci_min, ci_max, p, n_pairs, n)
            results.append((covariate, ndd, model, timeline, lag, HR, ci_min, ci_max, p, n_pairs, n))
            
cox1 = pd.DataFrame(results, columns=('PRIOR','OUTCOME', 'MODEL','TIMELINE', 'LAG', 'HR', 'ci_min', "ci_max", 'P_VAL', "N_pairs", "N"))

#Combine results
output = pd.concat([cox1])
output = output[output['N_pairs'] > 4]

#Adding FDR Correction

#Sort P-values
output = output.sort_values(by = "P_VAL")

#Drop Nan-values
output = output.dropna()

#FDR Correction
rejected, p_corr = fdrcorrection(output['P_VAL'], is_sorted=True)
output['P_CORR'] = p_corr
output['REJECTED'] = rejected

output

codes = pd.read_csv('files_for_cox/sleep_ICD10_codes_with_description.csv')
codes['DATE'] = codes['ICD10'] + '_DATE'
codes

final_results = output.merge(codes, left_on='PRIOR', right_on = 'DATE', how='left')
final_results['COHORT'] = 'UKB'
final_results = final_results.rename(columns = {'REJECTED':'SIGNIFICANT'})
final_results.loc[final_results.LAG == '0', 'TIMELINE'] = 'all'
final_results.loc[final_results.LAG == '0-1', 'TIMELINE'] = '< 1 year before tenure'
final_results.loc[final_results.LAG == '1-5', 'TIMELINE'] = '1-5 years before tenure'
final_results.loc[final_results.LAG == '5-10', 'TIMELINE'] = '5-10 years before tenure'
final_results.loc[final_results.LAG == '10-15', 'TIMELINE'] = '10-15 years before tenure'
final_results.loc[final_results.LAG == '5-15', 'TIMELINE'] = '5-15 years before tenure'
final_results = final_results[['ICD10', 'PRIOR', 'Description', 'OUTCOME', 'COHORT', 'LAG', 'TIMELINE', 'HR', 'ci_min', 'ci_max', 'P_VAL',
       'N_pairs', 'N', 'P_CORR', 'SIGNIFICANT']]
final_results = final_results.drop(columns = 'PRIOR')
final_results = final_results.rename(columns = {'Description':'PRIOR'})
final_results

final_results.to_csv(f'SLEEP_Supplementary_Table_3_45_no_wales.csv', header = True, index = False)

!dx upload SLEEP_Supplementary_Table_3_45_no_wales.csv --path /results/SLEEP_Supplementary_Table_3_45_no_wales.csv
