==> 09_PRS_density_plots.ipynb <==
# Data Prep

!pip install statsmodels
!pip install -U seaborn
!pip install lifelines
!pip install scikit-learn

# Imports here.
import numpy as np
import pandas as pd
import os
import seaborn as sns

import warnings
warnings.filterwarnings("ignore")

import scipy
from scipy import stats
from statsmodels.stats.weightstats import ztest as ztest

# Create density plot function

def density_plot(groups, column):

    # Iterate through the two groups
    for group in groups:
        # Subset
        subset = df_prs[df_prs[column] == group]

        # Draw the density plot
        plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                     kde_kws = {'linewidth': 3},
                     label = group)

    # Plot formatting
    plt.legend(prop={'size': 16}, title = column)
    plt.set_title(title)
    plt.set_xlabel(f'{ndd} PRS')
    plt.set_ylabel('Density')

# PD

# Set variables
ndd = 'PD'
df_prs = pd.read_csv(f'{ndd}_with_Z_score_for_graphs_april_30.csv')
df_prs

## Comparing PRS of cases and controls

groups = ['case', 'control']
column = 'status'
title = f'Case/Control Status for {ndd}'
density_plot(groups, column)

cases = df_prs[df_prs['status']=='case']
controls = df_prs[df_prs['status']=='control']

print(f'{ndd} Status ')
print("Cases:", len(cases))
print('Controls: ', len(controls))

print('T_test')
print(stats.ttest_ind(cases.Z_score, controls.Z_score))

## Comparing PRS of people with and without F51 Sleep Disorder

groups = ['no', 'yes']
column = 'F51 Disorder'
title = 'Comparing people with/without F51 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## Comparing PRS of people with/without G47 Sleep Disorder

groups = ['no', 'yes']
column = 'G47 Disorder'
title = 'Comparing people with/without G47 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## People with PD AND F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(7)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## People with PD AND G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(6)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

# AD - with APOE

# Set variables
ndd = 'AD'
df_prs = pd.read_csv(f'{ndd}_with_Z_score_for_graphs_april_30.csv')
df_prs

## Comparing PRS of cases and controls

groups = ['case', 'control']
column = 'status'
title = f'Case/Control Status for {ndd}'
density_plot(groups, column)

cases = df_prs[df_prs['status']=='case']
controls = df_prs[df_prs['status']=='control']

print(f'{ndd} Status ')
print("Cases:", len(cases))
print('Controls: ', len(controls))

print('T_test')
print(stats.ttest_ind(cases.Z_score, controls.Z_score))

## Comparing PRS of people with and without F51 Sleep Disorder

groups = ['no', 'yes']
column = 'F51 Disorder'
title = 'Comparing people with/without F51 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## Comparing PRS of people with/without G47 Sleep Disorder

groups = ['no', 'yes']
column = 'G47 Disorder'
title = 'Comparing people with/without G47 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## People with AD AND F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(7)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## People with AD AND G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(6)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} (including APOE) and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('Z_test')
print(ztest(with_ndd_sleep, without, value = 0))
print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('Z_test')
print(ztest(with_ndd_sleep, without, value = 0))
print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

# AD - without APOE

# Set variables
ndd = 'AD'
df_prs = pd.read_csv(f'{ndd}_NO_APOE_with_Z_score_for_graphs_april_30.csv')
df_prs

## Comparing PRS of cases and controls

groups = ['case', 'control']
column = 'status'
title = f'Case/Control Status for {ndd}'
density_plot(groups, column)

cases = df_prs[df_prs['status']=='case']
controls = df_prs[df_prs['status']=='control']

print(f'{ndd} Status ')
print("Cases:", len(cases))
print('Controls: ', len(controls))

print('T_test')
print(stats.ttest_ind(cases.Z_score, controls.Z_score))

## Comparing PRS of people with and without F51 Sleep Disorder

groups = ['no', 'yes']
column = 'F51 Disorder'
title = 'Comparing people with/without F51 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## Comparing PRS of people with/without G47 Sleep Disorder

groups = ['no', 'yes']
column = 'G47 Disorder'
title = 'Comparing people with/without G47 Sleep Disorder'
density_plot(groups, column)

with_disorder = df_prs[df_prs[column]=='yes']
without = df_prs[df_prs[column]=='no']

print(column)
print("People with disorder:", len(with_disorder))
print("People without disorder:", len(without))

print('T_test')
print(stats.ttest_ind(with_disorder.Z_score, without.Z_score))

## People with AD AND F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(7)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## People with AD AND G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='case']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
    subset1 = df_prs[df_prs['status']=='case']
    subset1 = subset1[subset1[disorder] == 'yes']

    subset2 = df_prs[df_prs['status']=='case']
    subset2 = subset2[subset2[disorder] == 'no']

    with_ndd_sleep = list(subset1[f'Z_score'])
    without = list(subset2[f'Z_score'])
    a = pd.DataFrame(stats.ttest_ind(with_ndd_sleep, without))
    s = a.loc[0,0].round(2)
    p = a.loc[1,0].round(6)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'People with {ndd} (excluding APOE) and {disorder} \n T-test: statistic {s}, p-value {p}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='case']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='case']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'{ndd} and {disorder}')
print(f"People with {ndd} and {disorder}:", len(with_ndd_sleep))
print(f"People with {ndd} and without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with F51 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'F51 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))

## Controls with G47 Disorder

# Cases vs. Controls
groups = ['no', 'yes']
disorder = 'G47 Disorder'

# Iterate through the two groups
for group in groups:
    # Subset
    subset = df_prs[df_prs['status']=='control']
    subset = subset[subset[disorder] == group]
    
    # Draw the density plot
    plt = sns.distplot(subset[f'Z_score'], hist = False, kde = True,
                 kde_kws = {'linewidth': 3},
                 label = group)
    
# Plot formatting
plt.legend(prop={'size': 16}, title = disorder)
plt.set_title(f'Controls and {disorder}')
plt.set_xlabel(f'{ndd} PRS')
plt.set_ylabel('Density')

subset1 = df_prs[df_prs['status']=='control']
subset1 = subset1[subset1[disorder] == 'yes']

subset2 = df_prs[df_prs['status']=='control']
subset2 = subset2[subset2[disorder] == 'no']

with_ndd_sleep = list(subset1[f'Z_score'])
without = list(subset2[f'Z_score'])

print(f'Controls with/without {disorder}')
print(f"Controls and {disorder}:", len(with_ndd_sleep))
print(f"Controls without {disorder}:", len(without))

print('T_test')
print(stats.ttest_ind(with_ndd_sleep, without))
